+++ TTPS/lab_admin/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from system_admin.models import Usuario
 from estudios.models import Estudio
+from django.core.validators import MinValueValidator
 
 # Create your models here.
 class Presupuesto(models.Model):
@@ -13,15 +14,29 @@ class Presupuesto(models.Model):
     class Meta:
         db_table = 'presupuestos'
 
+
+class Turno(models.Model):
+    id_turno = models.AutoField(primary_key=True)
+    usuario = models.ForeignKey(Usuario, on_delete=models.CASCADE)
+    estudio = models.ForeignKey(Estudio, on_delete=models.CASCADE)
+    fecha = models.DateField(null=True, blank=True)
+    numero = models.IntegerField(validators=[MinValueValidator(1)])
+
+    def __str__(self):
+        return f"Turno: {self.numero}"
+
+    class Meta:
+        db_table = 'turno'
+
 class Centro(models.Model):
     id_centro = models.AutoField(primary_key=True)
     nombre = models.ForeignKey(Estudio, on_delete=models.PROTECT)
 
     def __str__(self):
-        return f"Presupuesto #{self.id_presupuesto} - {self.estudio.id_interno}"
+        return f"Centro #{self.nombre}"
 
     class Meta:
-        db_table = 'presupuestos'
+        db_table = 'centro'
 
 class LabAdmin(models.Model):
     id_lab_admin = models.AutoField(primary_key=True)
